use fp32 for gradscaler

+++ pytorch-fsdp/fsdp_finetune.py
@@ -87,12 +87,12 @@ class Trainer:
         )
 
         # Initialize model with FSDP
-        model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, torch_dtype=torch.float16)
+        model = AutoModelForCausalLM.from_pretrained(MODEL_NAME)
         model.gradient_checkpointing_enable(gradient_checkpointing_kwargs={"use_reentrant": False})
         fsdp_model = FSDP(
             model,
             mixed_precision=MixedPrecision(
-                param_dtype=torch.float16,
+                param_dtype=torch.float32,
                 reduce_dtype=torch.float16,
                 buffer_dtype=torch.float16,
             ),
